maybe final

+++ react-admin/src/context/listContext/apiCalls.js
@@ -1,18 +1,25 @@
 import axios from 'axios';
 import { createListFailure, createListStart, createListSuccess, deleteListFailure, deleteListStart, deleteListSuccess, getListsFailure, getListsStart, getListsSuccess, updateListFailure, updateListStart, updateListSuccess } from './ListActions';
+import eventBus from '../../common/EventBus';
 
 //  GET LIST
 export const getLists = async (dispatch)=>{
     dispatch(getListsStart());
     try{
-        const res=await axios.get('lists/getAll',{
+        const res=await axios.get(process.env.REACT_APP_LINK+'lists/getAll',{
             headers:{
                 token:"Bearer "+ JSON.parse(localStorage.getItem("user")).accessToken
             }
         })
         dispatch(getListsSuccess(res.data));
+        return 'success'
     }catch(err){
-        dispatch(getListsFailure());
+        if(err && err.response.status===403){
+            return 403
+        }else{
+            dispatch(getListsFailure());
+            return 'failure'
+        }
     }
 }
 
@@ -20,14 +27,20 @@ export const getLists = async (dispatch)=>{
 export const deleteList = async (id,dispatch)=>{
     dispatch(deleteListStart());
     try{
-        await axios.delete('lists/'+id,{
+        await axios.delete(process.env.REACT_APP_LINK+'lists/'+id,{
             headers:{
                 token:"Bearer "+ JSON.parse(localStorage.getItem("user")).accessToken
             }
         })
         dispatch(deleteListSuccess(id));
+        return 'success'
     }catch(err){
-        dispatch(deleteListFailure());
+        if(err && err.response.status===403){
+            return 403;
+        }else{
+            dispatch(deleteListFailure());
+            return 'failure'
+        }
     }
 }
 
@@ -35,7 +48,7 @@ export const deleteList = async (id,dispatch)=>{
 export const createList= async (list,dispatch)=>{
     dispatch(createListStart());
     try{
-        const res=await axios.post('lists',list,{
+        const res=await axios.post(process.env.REACT_APP_LINK+'lists',list,{
             headers:{
                 token:"Bearer "+ JSON.parse(localStorage.getItem("user")).accessToken
             }
@@ -43,8 +56,12 @@ export const createList= async (list,dispatch)=>{
         dispatch(createListSuccess(res.data));
         return 'success'
     }catch(err){
+        if(err && err.response.status===403){
+            return 403
+        }else{
         dispatch(createListFailure());
         return 'failure'
+        }
     }
 }
 
@@ -52,7 +69,7 @@ export const createList= async (list,dispatch)=>{
 export const updateList= async (list,dispatch)=>{
     dispatch(updateListStart());
     try{
-        const res=await axios.put('/lists/'+list._id,list,{
+        const res=await axios.put(process.env.REACT_APP_LINK+'lists/'+list._id,list,{
             headers:{
                 token:"Bearer " + JSON.parse(localStorage.getItem("user")).accessToken
             }
@@ -60,7 +77,11 @@ export const updateList= async (list,dispatch)=>{
         dispatch(updateListSuccess(res.data));
         return 'success';
     }catch(err){
+        if(err && err.response.status===403){
+            return 403
+        }else{
         dispatch(updateListFailure());
         return 'failure';
+        }
     }
 }
